+++ BankAccount.h
@@ -3,22 +3,22 @@ using namespace std;
 
 class BankAccount {
 public:
-    float balance {0.00};
+    long balance {0};
     string name {};
     string username {};
     string password {};
 
     bool checkAccountExists(string username);
 
-    void createAccount(string username, string password, string name, float balance);
+    void createAccount(string username, string password, string name, long balance);
 
     bool logIn(string username, string password);
 
-    void addFunds(float depositedMoney);
+    void addFunds(long depositedMoney);
 
-    void withdrawFunds(float withdrawnMoney);
+    void withdrawFunds(long withdrawnMoney);
 
-    void transferMoney(string userToTransferTo, float moneyToTransfer);
+    void transferMoney(string userToTransferTo, long moneyToTransfer);
 
     void changeName(string newName);
